[PersistenceDiagramDistanceMatrix] Merge base module
