+++ core/base/persistenceDiagramDistanceMatrix/PersistenceDiagramDistanceMatrix.h
@@ -4,87 +4,152 @@
 /// \author Joseph Budin <joseph.budin@polytechnique.edu>
 /// \date September 2019
 ///
-/// \brief TTK processing package for the computation of Wasserstein barycenters
-/// and K-Means clusterings of a set of persistence diagrams.
-///
 /// \b Related \b publication \n
 /// "Progressive Wasserstein Barycenters of Persistence Diagrams" \n
 /// Jules Vidal, Joseph Budin and Julien Tierny \n
 /// Proc. of IEEE VIS 2019.\n
 /// IEEE Transactions on Visualization and Computer Graphics, 2019.
 ///
-/// \sa ttkPersistenceDiagramDistanceMatrix
+/// \sa PersistenceDiagramClustering
 
 #pragma once
 
-// base code includes
+#include <Auction.h>
+#include <KDTree.h>
 //
-#include <Wrapper.h>
-//
-#include <PersistenceDiagram.h>
-//
+#include <array>
 #include <limits>
-//
-#include <PDDistMat.h>
 //
 
 using namespace std;
-using namespace ttk;
 
 namespace ttk {
+
+  using DiagramTuple = std::tuple<ttk::SimplexId,
+                                  ttk::CriticalType,
+                                  ttk::SimplexId,
+                                  ttk::CriticalType,
+                                  double,
+                                  ttk::SimplexId,
+                                  double,
+                                  float,
+                                  float,
+                                  float,
+                                  double,
+                                  float,
+                                  float,
+                                  float>;
+  using MatchingTuple = std::tuple<ttk::SimplexId, ttk::SimplexId, double>;
+
   class PersistenceDiagramDistanceMatrix : public Debug {
 
   public:
-    PersistenceDiagramDistanceMatrix() {
-      wasserstein_ = 2;
-      use_progressive_ = 1;
-      use_kmeanspp_ = 0;
-      use_accelerated_ = 0;
-      numberOfInputs_ = 0;
-      threadNumber_ = 1;
-    };
+    std::vector<int>
+      execute(std::vector<std::vector<DiagramTuple>> &intermediateDiagrams);
 
-    ~PersistenceDiagramDistanceMatrix(){};
+    double getMostPersistent(int type = -1);
+    vector<vector<int>> get_centroids_sizes();
+    double getLessPersistent(int type = -1);
+    std::vector<std::vector<double>> getMinDiagonalPrices();
+    std::vector<std::vector<double>> getMinPrices();
 
-    void execute(std::vector<std::vector<DiagramTuple>> &intermediateDiagrams);
+    double computeDistance(const BidderDiagram<double> &D1,
+                           const BidderDiagram<double> &D2,
+                           const double delta_lim);
+    double computeDistance(const BidderDiagram<double> D1,
+                           const GoodDiagram<double> D2,
+                           const double delta_lim);
+    double computeDistance(BidderDiagram<double> *const D1,
+                           const GoodDiagram<double> *const D2,
+                           const double delta_lim);
+    double computeDistance(const GoodDiagram<double> &D1,
+                           const GoodDiagram<double> &D2,
+                           const double delta_lim);
+
+    GoodDiagram<double>
+      centroidWithZeroPrices(const GoodDiagram<double> centroid);
+    BidderDiagram<double> centroidToDiagram(const GoodDiagram<double> centroid);
+    GoodDiagram<double> diagramToCentroid(const BidderDiagram<double> diagram);
+    BidderDiagram<double>
+      diagramWithZeroPrices(const BidderDiagram<double> diagram);
+
+    void setBidderDiagrams();
+    void initializeEmptyClusters();
+    void initializeCentroids();
+    void initializeCentroidsKMeanspp();
+    void initializeAcceleratedKMeans();
+    void printDistancesToFile();
+    double computeRealCost();
+
+    std::vector<double> enrichCurrentBidderDiagrams(
+      std::vector<double> previous_min_persistence,
+      std::vector<double> min_persistence,
+      std::vector<std::vector<double>> initial_diagonal_prices,
+      std::vector<std::vector<double>> initial_off_diagonal_points,
+      std::vector<int> min_points_to_add,
+      bool add_points_to_barycenter);
+
+    std::vector<std::vector<double>> getDistanceMatrix();
+    void getCentroidDistanceMatrix();
+    void computeDiagramsDistanceMatrix();
+
+    void updateClusters();
+    void invertClusters();
+    void invertInverseClusters();
+
+    void acceleratedUpdateClusters();
+
+    inline void resetDosToOriginalValues() {
+      do_min_ = original_dos[0];
+      do_sad_ = original_dos[1];
+      do_max_ = original_dos[2];
+    }
 
     inline void setNumberOfInputs(int numberOfInputs) {
       numberOfInputs_ = numberOfInputs;
     }
 
-    inline void setWasserstein(const std::string &wasserstein) {
-      wasserstein_ = (wasserstein == "inf") ? -1 : stoi(wasserstein);
+    inline void setK(const int k) {
+      k_ = k;
     }
 
-    inline void setThreadNumber(const int &ThreadNumber) {
-      threadNumber_ = ThreadNumber;
+    inline void setWasserstein(const int &wasserstein) {
+      wasserstein_ = wasserstein;
+    }
+
+    inline void setThreadNumber(const int &threadNumber) {
+      threadNumber_ = threadNumber;
     }
 
     inline void setUseProgressive(const bool use_progressive) {
       use_progressive_ = use_progressive;
     }
 
-    inline void setAlpha(const double alpha) {
-      alpha_ = alpha;
+    inline void setKMeanspp(const bool use_kmeanspp) {
+      use_kmeanspp_ = use_kmeanspp;
     }
-    inline void setLambda(const double lambda) {
-      lambda_ = lambda;
+
+    inline void setUseKDTree(const bool use_kdtree) {
+      use_kdtree_ = use_kdtree;
+    }
+
+    inline void setAccelerated(const bool use_accelerated) {
+      use_accelerated_ = use_accelerated;
     }
 
     inline void setTimeLimit(const double time_limit) {
       time_limit_ = time_limit;
     }
 
-    inline void setUseKmeansppInit(const bool UseKmeansppInit) {
-      use_kmeanspp_ = UseKmeansppInit;
+    inline void setPairTypeClustering(const int pairTypeClustering) {
+      pairTypeClustering_ = pairTypeClustering;
     }
 
-    inline void setUseAccelerated(const bool UseAccelerated) {
-      use_accelerated_ = UseAccelerated;
+    inline void setGeometricalFactor(const double geometrical_factor) {
+      geometrical_factor_ = geometrical_factor;
     }
-
-    inline void setNumberOfClusters(const int NumberOfClusters) {
-      n_clusters_ = NumberOfClusters;
+    inline void setLambda(const double lambda) {
+      lambda_ = lambda;
     }
     inline void setForceUseOfAlgorithm(const bool forceUseOfAlgorithm) {
       forceUseOfAlgorithm_ = forceUseOfAlgorithm;
@@ -92,66 +157,143 @@ namespace ttk {
     inline void setDeterministic(const bool deterministic) {
       deterministic_ = deterministic;
     }
-    inline void setPairTypeClustering(const int pairTypeClustering) {
-      pairTypeClustering_ = pairTypeClustering;
-    }
 
-    inline void setUseDeltaLim(const bool useDeltaLim) {
-      useDeltaLim_ = useDeltaLim;
+    inline void setUseDeltaLim(const bool UseDeltaLim) {
+      UseDeltaLim_ = UseDeltaLim;
+      if(UseDeltaLim_) {
+        epsilon_min_ = 1e-8;
+      } else {
+        epsilon_min_ = 5e-5;
+      }
     }
 
     inline void setDistanceWritingOptions(const int distanceWritingOptions) {
       distanceWritingOptions_ = distanceWritingOptions;
     }
-
     inline void setDeltaLim(const double deltaLim) {
       deltaLim_ = deltaLim;
     }
+
     inline void setOutputDistanceMatrix(const bool arg) {
       outputDistanceMatrix_ = arg;
     }
     inline void setUseFullDiagrams(const bool arg) {
       useFullDiagrams_ = arg;
     }
-    inline void setPerClusterDistanceMatrix(const bool arg) {
-      perClusterDistanceMatrix_ = arg;
+
+    inline void printClustering() {
+      std::stringstream msg;
+      for(int c = 0; c < k_; ++c) {
+        msg << "[PersistenceDiagramClustering] Cluster " << c << " = {";
+        for(unsigned int idx = 0; idx < clustering_[c].size(); ++idx) {
+          if(idx == clustering_[c].size() - 1) {
+            msg << clustering_[c][idx] << "}" << std::endl;
+          } else {
+            msg << clustering_[c][idx] << ", ";
+          }
+        }
+      }
+      dMsg(std::cout, msg.str(), infoMsg);
     }
 
-    inline const std::vector<std::vector<double>> &&getDiagramsDistMat() {
-      return std::move(diagramsDistMat_);
+    inline void printOldClustering() {
+      std::stringstream msg;
+      for(int c = 0; c < k_; ++c) {
+        msg << "Cluster " << c << " = {";
+        for(unsigned int idx = 0; idx < old_clustering_[c].size(); ++idx) {
+          if(idx == old_clustering_[c].size() - 1) {
+            msg << old_clustering_[c][idx] << "}" << std::endl;
+          } else {
+            msg << old_clustering_[c][idx] << ", ";
+          }
+        }
+      }
+      dMsg(std::cout, msg.str(), infoMsg);
+    }
+
+    inline const std::vector<std::vector<double>> &&
+      getDiagramsDistanceMatrix() {
+      return std::move(diagramsDistanceMatrix_);
     }
 
   protected:
-    // Critical pairs used for clustering
-    // 0:min-saddles ; 1:saddles-saddles ; 2:sad-max ; else : all
-
+    bool barycenter_inputs_reset_flag;
+    bool precision_criterion_{false};
+    bool precision_max_{false};
+    bool precision_min_{false};
+    bool precision_sad_{false};
+    bool forceUseOfAlgorithm_{false};
+    bool deterministic_{true};
+    int wasserstein_{2};
+    double geometrical_factor_{1};
     double deltaLim_;
-    bool useDeltaLim_;
-    int distanceWritingOptions_;
-    int pairTypeClustering_;
-    bool forceUseOfAlgorithm_;
-    bool deterministic_;
-    int wasserstein_;
-    int n_clusters_;
+    bool UseDeltaLim_{false};
+    int distanceWritingOptions_{0};
+    // lambda : 0<=lambda<=1
+    // parametrizes the point used for the physical (critical) coordinates of
+    // the persistence paired lambda = 1 : extremum (min if pair min-sad, max if
+    // pair sad-max) lambda = 0 : saddle (bad stability) lambda = 1/2 : middle
+    // of the 2 critical points of the pair
+    double lambda_;
 
+    int k_;
     int numberOfInputs_;
-    int threadNumber_;
-    bool use_progressive_;
+    int threadNumber_{1};
+    bool use_progressive_{true};
     bool use_accelerated_;
     bool use_kmeanspp_;
-    double alpha_;
-    double lambda_;
-    double time_limit_;
+    bool use_kdtree_{true};
+    double time_limit_{std::numeric_limits<double>::max()};
 
-    int points_added_;
-    int points_deleted_;
+    double epsilon_min_{1e-8};
+    std::array<double, 3> epsilon_;
+    double cost_;
+    double cost_min_{0};
+    double cost_sad_{0};
+    double cost_max_{0};
 
-    std::vector<BidderDiagram<double>> bidder_diagrams_;
-    std::vector<GoodDiagram<double>> barycenter_goods_;
+    std::vector<std::vector<int>> current_bidder_ids_min_;
+    std::vector<std::vector<int>> current_bidder_ids_sad_;
+    std::vector<std::vector<int>> current_bidder_ids_max_;
+    std::vector<std::vector<DiagramTuple>> inputDiagramsMin_;
+    std::vector<std::vector<DiagramTuple>> inputDiagramsSaddle_;
+    std::vector<std::vector<DiagramTuple>> inputDiagramsMax_;
 
+    std::array<bool, 3> original_dos;
+
+    bool do_min_;
+    std::vector<BidderDiagram<double>> bidder_diagrams_min_;
+    std::vector<BidderDiagram<double>> current_bidder_diagrams_min_;
+    std::vector<GoodDiagram<double>> centroids_min_;
+    std::vector<GoodDiagram<double>> centroids_with_price_min_;
+
+    bool do_sad_;
+    std::vector<BidderDiagram<double>> bidder_diagrams_saddle_;
+    std::vector<BidderDiagram<double>> current_bidder_diagrams_saddle_;
+    std::vector<GoodDiagram<double>> centroids_saddle_;
+    std::vector<GoodDiagram<double>> centroids_with_price_saddle_;
+
+    bool do_max_;
+    std::vector<BidderDiagram<double>> bidder_diagrams_max_;
+    std::vector<BidderDiagram<double>> current_bidder_diagrams_max_;
+    std::vector<GoodDiagram<double>> centroids_max_;
+    std::vector<GoodDiagram<double>> centroids_with_price_max_;
+
+    std::vector<std::vector<int>> clustering_;
+    std::vector<std::vector<int>> old_clustering_;
+    std::vector<int> inv_clustering_;
+
+    std::vector<std::vector<int>> centroids_sizes_;
+
+    std::vector<bool> r_;
+    std::vector<double> u_;
+    std::vector<std::vector<double>> l_;
+    std::vector<std::vector<double>> centroidsDistanceMatrix_{};
+    std::vector<std::vector<double>> diagramsDistanceMatrix_{};
     bool outputDistanceMatrix_{false};
     bool useFullDiagrams_{false};
-    bool perClusterDistanceMatrix_{false};
-    std::vector<std::vector<double>> diagramsDistMat_{};
+
+    int n_iterations_;
+    int pairTypeClustering_;
   };
 } // namespace ttk
